Replaced "1 minute" with "6 minutes"
Since removing the bottom cover is required in order to replace the battery and removing the bottom cover takes 5 minutes, the time it takes to battery must take more than 5 minutes. It said 1 minute before, so I just added 1 minute to the 5 minutes it takes to remove the bottom cover.