Add Reese's Law small battery warning to all sections
Adds ingestion hazard warning to all sections including mention of a CMOS battery in order to comply with [Reese's Law](https://www.congress.gov/bill/117th-congress/house-bill/5313). This exact verbiage is not required and can be tweaked if desired in the future, but it was drafted based on documentation from the [Federal Register](https://www.federalregister.gov/documents/2023/09/21/2023-20334/safety-standard-for-button-cell-or-coin-batteries-and-consumer-products-containing-such-batteries) and the [Consumer Product Safety Commission](https://www.cpsc.gov/Business--Manufacturing/Business-Education/Business-Guidance/Button-Cell-and-Coin-Battery).