bonw16 & serw14: Specs
The marketing team would prefer we add placeholder sections so they can set the links up at the same time as the rest of the catalog details and don't need to monitor when tech docs become available. This PR adds placeholders for bonw16 and serw14 to fix the current broken links.